+++ RELEASE_NOTES.md
@@ -6,6 +6,19 @@ Until then, breaking changes can happen in minor versions.
 
 
 
+## 0.2
+
+### ✨ Features
+
+* ``ColorPickerBar`` takes bar colors, reset value and supports opacity directly.
+* ``ColorPickerBar.Style`` is the new owner of the `resetButton` property.
+
+### 🗑️ Deprecations
+
+* ``ColorPickerBar.Configuration`` is deprecated.
+
+
+
 ## 0.1.1
 
 ### ✨ Features
@@ -23,6 +36,6 @@ This is the first public release of PickerKit.
 * ``Camera`` can be used to take photos and handle them as images.
 * ``ColorPickerBar`` adds a color picker to a horizontal or vertical bar with additional colors.
 * ``DocumentScanner`` can be used to scan documents and handle them as images.
-* ``FilePicker`` can be used to pick any file from the Files app.
-* ``ImagePicker`` can be used to pick images from the user's photo library.
+* ``FilePicker`` can be used to pick files from the Files app.
+* ``ImagePicker`` can be used to pick images from the  photo library.
 * ``MultiPicker`` can be used to pick multiple items in e.g. a list or form.